- Renumber Photobank error codes from 24XX to 26XX (24XX claimed by Articles, 25XX by Leaflet)
- Keep both photobank routes and leaflet-generator route in App.tsx
- Keep photobank sidebar entry alongside leaflet/articles/knowledge-base entries
- Keep photobank and article generation i18n error messages

+++ backend/src/Adapters/Anela.Heblo.Adapters.Flexi/Manufacture/FlexiManufactureClient.cs
@@ -132,7 +132,7 @@ internal class FlexiManufactureClient : IManufactureClient
                             ProductCode = existing.ProductCode,
                             ProductName = existing.ProductName,
                             ProductType = existing.ProductType,
-                            RequiredAmount = (double)distributionEntry.AdjustedConsumption,
+                            RequiredAmount = distributionEntry.AdjustedConsumption,
                             HasLots = existing.HasLots
                         };
                     }
